+++ src/redux/store.js
@@ -1,4 +1,4 @@
-import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
+import { configureStore,  getDefaultMiddleware } from '@reduxjs/toolkit';
 import {
   persistStore,
   persistReducer,
@@ -37,3 +37,6 @@ export const store = configureStore({
 });
 
 export const persistor = persistStore(store);
+
+
+
