+++ src/redux/transaction/transaction-operation.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { createAsyncThunk } from '@reduxjs/toolkit';
+import { createAsyncThunk, createAction } from '@reduxjs/toolkit';
 
 const getAllTransactions = createAsyncThunk('transactions/getAll', async () => {
   try {
@@ -61,13 +61,15 @@ const getCategory = createAsyncThunk(
   async (_, { rejectWithValue }) => {
     try {
       const { data: response } = await axios.get('/categories');
-      console.log(response);
       return response;
     } catch (err) {
       return rejectWithValue(err.response.data);
     }
   }
 );
+
+ const  ModalAddTrans = createAction('transactions/ModalAddTrans');
+
 
 // const updateTransactions = createAsyncThunk(
 //   'transactions/update',
@@ -90,6 +92,7 @@ const transactionsOperation = {
   deleteTransactions,
   getByDate,
   getCategory,
+  ModalAddTrans
 };
 
 export default transactionsOperation;
