+++ src/redux/transaction/transaction-selectors.js
@@ -1,7 +1,7 @@
 const getTransaction = state => state.transaction.items;
 const isLoading = state => state.transaction.isLoading;
 const isModalAddTransOpen = state => state.transaction.isModalAddTransOpen;
-const getCategories = state => state.categories;
+const getCategories = state => state.transaction.categories;
 const getStatistics = state => state.transaction;
 
 const transactionSelector = {
