+++ src/redux/transaction/transaction-slice.js
@@ -1,4 +1,4 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice} from '@reduxjs/toolkit';
 import transactionOperation from './transaction-operation';
 
 const initialState = {
@@ -6,11 +6,17 @@ const initialState = {
   isLoading: true,
   categories: {},
   totalTransactions: {},
+  isModalAddTransOpen: false
 };
 
 const transactionsSlice = createSlice({
   name: 'transactions',
   initialState,
+  reducers: {
+     ModalAddTrans(state, action) {
+      state.isModalAddTransOpen = action.payload;
+    },
+  },
 
   extraReducers: {
     [transactionOperation.getAllTransactions.fulfilled]: (state, action) => {
